+++ android/src/main/java/com/appboy/reactbridge/AppboyReactPackage.java
@@ -22,7 +22,7 @@ public class AppboyReactPackage implements ReactPackage {
     return modules;
   }
 
-  // This "Override" method is being kept for backwards compatibilty with React Native
+  // This "Override" method is being kept for backwards compatibility with React Native
   // releases earlier than https://github.com/facebook/react-native/releases/tag/v0.47.0
   public List<Class<? extends JavaScriptModule>> createJSModules() {
     return Collections.emptyList();